fix docs: `keyboardVerticalOffset`Not Required (#322)
Docs state that the `keyboardVerticalOffset` prop for `keyboardAvoidingView` is required.   
But I do not believe this to be the case.   
I have yet to see a working example where this prop is even supplied.   
Indeed, this component works for me on both ios and android, without supplying it..   
(And when I did briefly try it out, I did not notice any difference.)

I changed the docs from `keyboardVerticalOffset` `Required: Yes` to `Required: No`