Fixed the Function Component Example (#2019)
* Fixed the Function Component Example

We experienced that the documentation was faulty so after we at @enpit found a fix, we wanted to share it with you. 
The old trigger `appState.current.match(/inactive|background/) && nextAppState === "active"` would never work because the useState hook doesn't update the state fast enough for the function to work. The solution is to use useRef instead and optional (to display the inactive status under iOS) also use a `useState` for cosmetics.

* removed block comment, removed spaces

* rephrase and move new comment, add one more sentence, prettier run

* language lint fix

Co-authored-by: Bartosz Kaszubowski <gosimek@gmail.com>